Change flow from QRActivity to HomeActivity
