+++ app/src/main/java/com/alphawallet/app/C.java
@@ -82,7 +82,7 @@ public abstract class C {
     public static final String EXTRA_LOCALE = "LOCALE_STRING";
     public static final String EXTRA_PAGE_TITLE = "PTITLE";
     public static final String EXTRA_CURRENCY = "CURRENCY_STRING";
-    public static final String EXTRA_URL = "URL";
+    public static final String EXTRA_QR_CODE = "QR_SCAN_CODE";
     public static final String EXTRA_UNIVERSAL_SCAN = "UNIVERSAL_SCAN";
 
     public static final String PRUNE_ACTIVITY =
